# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateLevelRequest:
    properties:
      bearing:
        type: number
      elevation:
        type: integer
      external_id:
        type: string
      file_ids:
        type: string
      height:
        type: integer
      is_published:
        type: boolean
      latitude:
        type: number
      level_height:
        type: number
      level_width:
        type: number
      longitude:
        type: number
      map_group_id:
        type: string
      name:
        type: string
      scale:
        type: number
      short_name:
        type: string
      type:
        type: integer
      width:
        type: integer
    type: object
...
  dto.GeoReferenceRequest:
    properties:
      control_x:
        type: integer
      control_y:
        type: integer
      target_x:
        type: number
      target_y:
        type: number
    type: object
  dto.GeoReferenceResponse:
    properties:
      control_x:
        type: integer
      control_y:
        type: integer
      created_at:
        type: string
      id:
        type: string
      level_id:
        type: string
      target_x:
        type: number
      target_y:
        type: number
    type: object
...
  dto.LevelResponse:
    properties:
      bearing:
        type: number
      created_at:
        type: string
      elevation:
        type: integer
      external_id:
        type: string
      file_ids:
        type: string
      height:
        type: integer
      id:
        type: string
      is_published:
        type: boolean
      latitude:
        type: number
      level_height:
        type: number
      level_width:
        type: number
      longitude:
        type: number
      map_group_id:
        type: string
      name:
        type: string
      perspective:
        $ref: '#/definitions/dto.PerspectiveResponse'
      perspective_id:
        type: string
      scale:
        type: number
      short_name:
        type: string
      type:
        type: integer
      updated_at:
        type: string
      venue_id:
        type: string
      width:
        type: integer
    type: object
...
  dto.MapGroupResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      short_name:
        type: string
      sort_index:
        type: integer
      type:
        type: string
      updated_at:
        type: string
      venue_id:
        type: string
    type: object
...
  dto.PerspectiveRequest:
    properties:
      camera_max_zoom:
        type: number
      camera_min_zoom:
        type: number
      camera_target_bearing:
        type: number
      camera_target_center_lat:
        type: number
      camera_target_center_lng:
        type: number
      camera_target_pitch:
        type: number
      camera_target_zoom:
        type: number
      camera_type:
        type: integer
      camera_zoom:
        type: number
      name:
        type: string
    type: object
  dto.PerspectiveResponse:
    properties:
      camera_max_zoom:
        type: number
      camera_min_zoom:
        type: number
      camera_target_bearing:
        type: number
      camera_target_center_lat:
        type: number
      camera_target_center_lng:
        type: number
      camera_target_pitch:
        type: number
      camera_target_zoom:
        type: number
      camera_type:
        type: integer
      camera_zoom:
        type: number
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.Response:
    properties:
      code:
        $ref: '#/definitions/dto.AppCode'
      data: {}
      messages:
        items:
          type: string
        type: array
    type: object
...
  dto.UpdateLevelRequest:
    properties:
      bearing:
        type: number
      elevation:
        type: integer
      external_id:
        type: string
      file_ids:
        type: string
      height:
        type: integer
      is_published:
        type: boolean
      latitude:
        type: number
      level_height:
        type: number
      level_width:
        type: number
      longitude:
        type: number
      map_group_id:
        type: string
      name:
        type: string
      scale:
        type: number
      short_name:
        type: string
      type:
        type: integer
      width:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Digimap backend REST API
  title: Digimap Backend API
  version: "1.0"
paths:
...
  /venues/{id}/levels:
    get:
      description: List all levels for a venue
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.LevelResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: List levels
      tags:
      - levels
    post:
      consumes:
      - application/json
      description: Create a new level for a venue (requires editor role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Level details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateLevelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LevelResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Create level
      tags:
      - levels
  /venues/{id}/levels/{levelID}:
    delete:
      description: Delete a level (requires editor role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Level ID
        in: path
        name: levelID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Delete level
      tags:
      - levels
    get:
      description: Get a level by ID
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Level ID
        in: path
        name: levelID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LevelResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Get level
      tags:
      - levels
    put:
      consumes:
      - application/json
      description: Update a level (requires editor role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Level ID
        in: path
        name: levelID
        required: true
        type: string
      - description: Level details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateLevelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LevelResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Update level
      tags:
      - levels
  /venues/{id}/levels/{levelID}/geo-references:
    get:
      description: List geo references for a level
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Level ID
        in: path
        name: levelID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.GeoReferenceResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: List geo references
      tags:
      - levels
    post:
      consumes:
      - application/json
      description: Create a geo reference for a level (requires editor role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Level ID
        in: path
        name: levelID
        required: true
        type: string
      - description: Geo reference details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.GeoReferenceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.GeoReferenceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Create geo reference
      tags:
      - levels
  /venues/{id}/levels/{levelID}/geo-references/{refID}:
    delete:
      description: Delete a geo reference (requires editor role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Level ID
        in: path
        name: levelID
        required: true
        type: string
      - description: Geo Reference ID
        in: path
        name: refID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Delete geo reference
      tags:
      - levels
  /venues/{id}/levels/{levelID}/perspective:
    put:
      consumes:
      - application/json
      description: Create or update the camera perspective for a level (requires editor
        role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Level ID
        in: path
        name: levelID
        required: true
        type: string
      - description: Perspective details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.PerspectiveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PerspectiveResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Upsert perspective
      tags:
      - levels
...
  /venues/{id}/map-groups:
    get:
      description: List all map groups for a venue
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.MapGroupResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: List map groups
      tags:
      - levels
    post:
      consumes:
      - application/json
      description: Create a new map group for a venue (requires editor role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Map group details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MapGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MapGroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Create map group
      tags:
      - levels
  /venues/{id}/map-groups/{mgID}:
    delete:
      description: Delete a map group (requires editor role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Map Group ID
        in: path
        name: mgID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Delete map group
      tags:
      - levels
    put:
      consumes:
      - application/json
      description: Update a map group (requires editor role)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Map Group ID
        in: path
        name: mgID
        required: true
        type: string
      - description: Map group details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MapGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MapGroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Update map group
      tags:
      - levels
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"